+++ src/libs/blueprint/conduit_blueprint_mesh_matset_xforms.cpp
@@ -502,6 +502,11 @@ to_silo(const conduit::Node &field,
             slot_index += elem_mat_map.size();
         }
     }
+
+    // extra hooks for downstream data consumers
+
+    dest["buffer_style"] = mesh::matset::is_multi_buffer(matset) ? "multi" : "uni";
+    dest["dominance"] = mesh::matset::is_element_dominant(matset) ? "element" : "material";
 }
 
 //-----------------------------------------------------------------------------
@@ -1450,14 +1455,14 @@ to_multi_buffer_full(const conduit::Node &src_matset,
 
 //-----------------------------------------------------------------------------
 void
-to_sparse_by_element(const conduit::Node &src_matset,
-                     conduit::Node &dest_matset,
-                     const float64 epsilon)
+to_uni_buffer_by_element(const conduit::Node &src_matset,
+                         conduit::Node &dest_matset,
+                         const float64 epsilon)
 {
     // extra seat belt here
     if (! src_matset.dtype().is_object())
     {
-        CONDUIT_ERROR("blueprint::mesh::matset::to_sparse_by_element"
+        CONDUIT_ERROR("blueprint::mesh::matset::to_uni_buffer_by_element"
                       " passed matset node must be a valid matset tree.");
     }
 
@@ -1543,13 +1548,13 @@ to_silo(const conduit::Node &specset,
         const conduit::Node &matset,
         conduit::Node &dest)
 {
-    if(!specset.dtype().is_object() )
+    if(! specset.dtype().is_object())
     {
         CONDUIT_ERROR("blueprint::mesh::specset::to_silo passed specset node "
                       "must be a valid specset tree.");
     }
 
-    if(!matset.dtype().is_object() )
+    if(! matset.dtype().is_object())
     {
         CONDUIT_ERROR("blueprint::mesh::specset::to_silo passed matset node "
                       "must be a valid matset tree or a valid intermediate silo "
@@ -1564,7 +1569,9 @@ to_silo(const conduit::Node &specset,
            matset.has_child("mix_vf") && 
            matset.has_child("mix_mat") &&
            matset.has_child("mix_next") &&
-           matset.has_child("matlist")))
+           matset.has_child("matlist") &&
+           matset.has_child("buffer_style") &&
+           matset.has_child("dominance")))
     {
         // if not, create a silo rep
         conduit::blueprint::mesh::matset::to_silo(matset, silo_matset);
@@ -1575,54 +1582,81 @@ to_silo(const conduit::Node &specset,
         silo_matset.set_external(matset);
     }
 
-    const int nmat = silo_matset["material_map"].number_of_children();
-    CONDUIT_ASSERT(nmat == specset["matset_values"].number_of_children(),
-        "blueprint::mesh::specset::to_silo number of materials must match "
-        "between passed specset and passed matset.");
+    // TODO change this once we write specset converters
+    // I think the right path will be to rewrite this function using the 
+    // sparse by element (uni_buffer element_dominant) specset flavor.
+    // So we will convert all specsets to that form and then convert to silo.
+    // Should be simpler and get rid of a lot of the indexing madness.
+    if (silo_matset["buffer_style"].as_string() != "multi")
+    {
+        CONDUIT_ERROR("TODO cannot handle uni buffer specsets");
+    }
+    if (silo_matset["dominance"].as_string() != "element")
+    {
+        CONDUIT_ERROR("TODO cannot handle material dominant specsets");
+    }
 
-    auto matset_vals_itr = specset["matset_values"].children();
+    const int nmat = silo_matset["material_map"].number_of_children();
+    CONDUIT_ASSERT(nmat >= specset["matset_values"].number_of_children(),
+        "blueprint::mesh::specset::to_silo number of materials in the matset "
+        "must be greater than or equal to the number of materials in the specset.");
+
     auto matmap_itr = silo_matset["material_map"].children();
     int matmap_index = 0;
     // Map actual material numbers to indicies into the material map
     // We need this map so that, no matter what material numbers we get thrown at us,
     // we can figure out their order in the material map for when we calculate
     // species indices.
-    std::map<int, int> matmap_map;
-    while (matset_vals_itr.has_next() && matmap_itr.has_next())
+    std::map<int, int> mat_id_to_array_index;
+    while (matmap_itr.has_next())
     {
-        matset_vals_itr.next();
         const Node &matmap_entry = matmap_itr.next();
-        // Make sure that materials are in the same order across the specset and matset
-        CONDUIT_ASSERT(matset_vals_itr.name() == matmap_itr.name(), 
-            "blueprint::mesh::specset::to_silo materials must be in the same order "
-            "between passed specset and passed matset.");
-        matmap_map[matmap_entry.as_int()] = matmap_index;
+        mat_id_to_array_index[matmap_entry.as_int()] = matmap_index;
         matmap_index ++;
     }
 
+    //
+    // set nmatspec and specnames arrays
+    //
     dest["nmatspec"].set(DataType::index_t(nmat));
     index_t_array nmatspec = dest["nmatspec"].value();
-    matset_vals_itr.to_front();
-    int material_index = 0;
-    while (matset_vals_itr.has_next())
+    // we have to be very careful to always go in the order of the material map
+    int matmap_idx = 0;
+    matmap_itr.to_front();
+    while (matmap_itr.has_next())
     {
-        // get the number of species for this material
-        const Node &individual_mat_spec = matset_vals_itr.next();
-        const std::string matname = matset_vals_itr.name();
-        const int num_species_for_this_material = individual_mat_spec.number_of_children();
-        nmatspec[material_index] = num_species_for_this_material;
+        matmap_itr.next();
+        const std::string matname = matmap_itr.name();
 
-        // get the specie names for this material
-        auto spec_itr = individual_mat_spec.children();
-        while (spec_itr.has_next())
+        // is this material present in the specset?
+        if (specset["matset_values"].has_child(matname))
         {
-            spec_itr.next();
-            const std::string specname = spec_itr.name();
-            Node &specname_entry = dest["specnames"].append();
-            specname_entry.set(specname);
+            const Node &individual_mat_spec = specset["matset_values"][matname];
+            // get the number of species for this material
+            const int num_species_for_this_material = individual_mat_spec.number_of_children();
+            // save the number of species for this material in the output
+            nmatspec[matmap_idx] = num_species_for_this_material;
+
+            // get the specie names for this material and add to the specnames.
+            // the specnames array is the length of the sum of the nmatspec array
+            // so for all materials with species, the species names will appear
+            // in this list in order.
+            auto spec_itr = individual_mat_spec.children();
+            while (spec_itr.has_next())
+            {
+                spec_itr.next();
+                const std::string specname = spec_itr.name();
+                Node &specname_entry = dest["specnames"].append();
+                specname_entry.set(specname);
+            }
+        }
+        else
+        {
+            // if this material has no species, then we set to zero.
+            nmatspec[matmap_idx] = 0;
         }
 
-        material_index ++;
+        matmap_idx ++;
     }
 
     // we sum up the nmatspec to get the number of species across all materials
@@ -1631,7 +1665,7 @@ to_silo(const conduit::Node &specset,
     // we have to go in order by zones as they appear
 
     // first we need number of zones
-    const int nzones = silo_matset["matlist"].dtype().number_of_elements();
+    const int num_zones = silo_matset["matlist"].dtype().number_of_elements();
 
     // TODO
     // I may wish to go through and check if the material is even in the zone
@@ -1644,21 +1678,31 @@ to_silo(const conduit::Node &specset,
     std::vector<float64> species_mf;
     
     // need to iterate across all species for all materials at once
-    for (int zoneId = 0; zoneId < nzones; zoneId ++)
+    for (int zone_id = 0; zone_id < num_zones; zone_id ++)
     {
-        // iterate through each material
-        matset_vals_itr.to_front();
-        while (matset_vals_itr.has_next())
+        // we must iterate using the material map since it has the "correct"
+        // ordering of materials. Ordering may be different for the specset.
+        // We choose the material map order to be the one source of truth.
+        matmap_itr.to_front();
+        while (matmap_itr.has_next())
         {
-            const Node &individual_mat_spec = matset_vals_itr.next();
-            // iterate through each specie
-            auto spec_itr = individual_mat_spec.children();
-            while (spec_itr.has_next())
+            matmap_itr.next();
+            const std::string matname = matmap_itr.name();
+
+            // is this material present in the specset?
+            if (specset["matset_values"].has_child(matname))
             {
-                const Node &spec = spec_itr.next();
-                float64_accessor species_mass_fractions = spec.value();
-                // grab the specie mass fraction for this zone id
-                species_mf.push_back(species_mass_fractions[zoneId]);
+                // if so, we just load all the species mass fractions in
+                const Node &individual_mat_spec = specset["matset_values"][matname];
+                // iterate through each specie
+                auto spec_itr = individual_mat_spec.children();
+                while (spec_itr.has_next())
+                {
+                    const Node &spec = spec_itr.next();
+                    float64_accessor species_mass_fractions = spec.value();
+                    // grab the specie mass fraction for this zone id
+                    species_mf.push_back(species_mass_fractions[zone_id]);
+                }
             }
         }
     }
@@ -1670,7 +1714,7 @@ to_silo(const conduit::Node &specset,
     const int_accessor silo_mix_mat = silo_matset["mix_mat"].value();
     const int_accessor silo_mix_next = silo_matset["mix_next"].value();
 
-    auto calculate_species_index = [&](const int zoneId, const int mat_index)
+    auto calculate_species_index = [&](const int zone_id, const int mat_index)
     {
         // To get the value for the speclist for this zone, we must determine
         // the correct 1-index in the species_mf array that corresponds to the 
@@ -1683,7 +1727,7 @@ to_silo(const conduit::Node &specset,
         // array.
 
         // how many entries per zone? Use the calculated num_species_across_mats
-        const int outer_index = zoneId * num_species_across_mats;
+        const int outer_index = zone_id * num_species_across_mats;
 
         // Next we need the inner or "local_index", which corresponds to the 
         // starting 1-index of the relevant material's species within this zone.
@@ -1711,24 +1755,24 @@ to_silo(const conduit::Node &specset,
         // we save the final index for this zone
         return outer_index + local_index;
 
-        // TODO return a zero here if the material in the zone contains only 1 species.
-        // This is a further optimization. It doesn't matter for now since we are
-        // treating our output like we have all species and all materials in all zones.
-        // I think the point of this optimization is to also leave stuff out of the 
-        // species_mf array. If I want to do this then I should explore that as well.
+        // This can produce an out of bounds index in very specific cases.
+        // If a material has no species, the index produced by this function is 
+        // useless, but downstream data consumers shouldn't be reading the index
+        // anyway. If a material has no species and it is the last one in the 
+        // material map and the final zone is mixed and contains that material,
+        // then we can get an index that is out of bounds. This is ok because 
+        // downstream tools like VisIt read based on the number of species, so
+        // even though the index is garbage it goes unused.
     };
 
-    // our negative 1-index into the mix_spec array
-    int mix_start_index = -1;
-
-    dest["speclist"].set(DataType::int64(nzones));
+    dest["speclist"].set(DataType::int64(num_zones));
     int64_array speclist = dest["speclist"].value();
     std::vector<int> mix_spec;
 
     // now we create the speclist and mix_spec arrays, traversing through the zones
-    for (int zoneId = 0; zoneId < nzones; zoneId ++)
+    for (int zone_id = 0; zone_id < num_zones; zone_id ++)
     {
-        const int matlist_entry = silo_matlist[zoneId];
+        const int matlist_entry = silo_matlist[zone_id];
         // is this zone clean?
         if (matlist_entry >= 0) // this relies on matset_ptr->allowmat0 == 0
         {
@@ -1736,17 +1780,35 @@ to_silo(const conduit::Node &specset,
 
             // I can use the material number to determine which part of the speclist to index into
             const int &matno = matlist_entry;
-            const int mat_index = matmap_map[matno];
-            speclist[zoneId] = calculate_species_index(zoneId, mat_index);
+            const int mat_index = mat_id_to_array_index[matno];
+            if (nmatspec[mat_index] == 1)
+            {
+                // This is an optimization for if the material has only one
+                // species. See MIR.C in VisIt in the MIR::SpeciesSelect() 
+                // function to see how this optimization is used.
+                speclist[zone_id] = 0;
+            }
+            else
+            {
+                // Either there are multiple species for this material or there 
+                // are none. If there are none, then the value computed here
+                // will ultimately not be used by Silo readers. There must be 
+                // a value here though even when there are no species for the
+                // material because we must have entries in the different silo
+                // species arrays for each material.
+                speclist[zone_id] = calculate_species_index(zone_id, mat_index);
+            }
         }
         else
         {
             // mixed
 
-            // we save the negated 1-index into the mix_spec array
-            speclist[zoneId] = mix_start_index;
+            // We don't need to compute this as it is the same as the 
+            // matlist entry.
+            // We save the negated 1-index into the mix_spec array
+            speclist[zone_id] = matlist_entry;
 
-            // for mixed zones, the numbers in the matlist are negated 1-indices into
+            // for mixed zones, the numbers in the speclist are negated 1-indices into
             // the silo mixed data arrays. To turn them into zero-indices, we must add
             // 1 and negate the result. Example:
             // indices: -1 -2 -3 -4 ...
@@ -1759,17 +1821,30 @@ to_silo(const conduit::Node &specset,
             {                
                 // I can use the material number to determine which part of the speclist to index into
                 const int matno = silo_mix_mat[mix_id];
-                const int mat_index = matmap_map[matno];
-                mix_spec.push_back(calculate_species_index(zoneId, mat_index));
+                const int mat_index = mat_id_to_array_index[matno];
+                if (nmatspec[mat_index] == 1)
+                {
+                    // This is an optimization for if the material has only one
+                    // species. See MIR.C in VisIt in the MIR::SpeciesSelect() 
+                    // function to see how this optimization is used.
+                    mix_spec.push_back(0);
+                }
+                else
+                {
+                    // Either there are multiple species for this material or there 
+                    // are none. If there are none, then the value computed here
+                    // will ultimately not be used by Silo readers. There must be 
+                    // a value here though even when there are no species for the
+                    // material because we must have entries in the different silo
+                    // species arrays for each material.
+                    mix_spec.push_back(calculate_species_index(zone_id, mat_index));
+                }
 
                 // since mix_id is a 1-index, we must subtract one
                 // this makes sure that mix_id = 0 is the last case,
                 // since it will make our mix_id == -1, which ends
                 // the while loop.
                 mix_id = silo_mix_next[mix_id] - 1;
-
-                // decrement this index every time we write another index to the mix_spec
-                mix_start_index --;
             }
         }
     }
@@ -1781,10 +1856,10 @@ to_silo(const conduit::Node &specset,
     dest["nmat"] = nmat;
     
     // number of species associated with each material
-    // we already saved nmatspec
+    // we already saved dest["nmatspec"]
     
     // indices into species_mf and mix_spec
-    // we already saved speclist
+    // we already saved dest["speclist"]
     
     // length of the species_mf array
     dest["nspecies_mf"] = nspecies_mf;
@@ -1892,22 +1967,22 @@ to_multi_buffer_full(const conduit::Node &src_matset,
 
 //-----------------------------------------------------------------------------
 void
-to_sparse_by_element(const conduit::Node &src_matset,
-                     const conduit::Node &src_field,
-                     const std::string &dest_matset_name,
-                     conduit::Node &dest_field,
-                     const float64 epsilon)
+to_uni_buffer_by_element(const conduit::Node &src_matset,
+                         const conduit::Node &src_field,
+                         const std::string &dest_matset_name,
+                         conduit::Node &dest_field,
+                         const float64 epsilon)
 {
     // extra seat belt here
     if (! src_matset.dtype().is_object())
     {
-        CONDUIT_ERROR("blueprint::mesh::field::to_sparse_by_element"
+        CONDUIT_ERROR("blueprint::mesh::field::to_uni_buffer_by_element"
                       " passed matset node must be a valid matset tree.");
     }
 
     if (! src_field.dtype().is_object())
     {
-        CONDUIT_ERROR("blueprint::mesh::field::to_sparse_by_element"
+        CONDUIT_ERROR("blueprint::mesh::field::to_uni_buffer_by_element"
                       " passed field node must be a valid field tree.");
     }